# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "Basis": {
        "description": "How a claim is known. Independent of ``Confidence``.",
        "enum": [
          "observed",
          "resolved",
          "heuristic",
          "inferred",
          "interpreted"
        ],
        "title": "Basis",
        "type": "string"
      },
      "CallKind": {
        "enum": [
          "direct",
          "method",
          "constructor"
        ],
        "title": "CallKind",
        "type": "string"
      },
      "CallsAttributes": {
        "additionalProperties": false,
        "properties": {
          "call_kind": {
            "$ref": "#/components/schemas/CallKind"
          },
          "target_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Symbol"
          }
        },
        "required": [
          "call_kind"
        ],
        "title": "CallsAttributes",
        "type": "object"
      },
...
      "ClassificationFacet": {
        "enum": [
          "role",
          "origin",
          "layer"
        ],
        "title": "ClassificationFacet",
        "type": "string"
      },
      "Confidence": {
        "enum": [
          "high",
          "medium",
          "low"
        ],
        "title": "Confidence",
        "type": "string"
      },
      "ConfigEvidence": {
        "additionalProperties": false,
        "description": "A value in a structured config file (JSON, TOML, YAML).\n\n``json_pointer`` (RFC 6901) addresses the parsed data model, whatever the\nfile's syntax: ``/project/dependencies/0`` in a pyproject.toml.",
        "properties": {
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "file_key": {
            "title": "File Key",
            "type": "string"
          },
          "json_pointer": {
            "title": "Json Pointer",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "type": {
            "const": "config",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "basis",
          "provenance",
          "type",
          "file_key",
          "json_pointer"
        ],
        "title": "ConfigEvidence",
        "type": "object"
      },
      "ConventionEvidence": {
        "additionalProperties": false,
        "description": "A naming or layout convention matched, e.g. ``test_*.py`` on a file name.\n\nThe convention's identity is ``provenance.rule`` (required); ``pattern`` is\nits human-readable form, ``matched_value`` what it matched.",
        "properties": {
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "matched_value": {
            "title": "Matched Value",
            "type": "string"
          },
          "pattern": {
            "title": "Pattern",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "type": {
            "const": "convention",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "basis",
          "provenance",
          "type",
          "pattern",
          "matched_value"
        ],
        "title": "ConventionEvidence",
        "type": "object"
      },
...
      "DerivedEvidence": {
        "additionalProperties": false,
        "description": "Computed from other records of the same snapshot (rollups, inference).",
        "properties": {
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "edge_refs": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EdgeRef"
            },
            "title": "Edge Refs",
            "type": "array"
          },
          "node_keys": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Node Keys",
            "type": "array"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "type": {
            "const": "derived",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "basis",
          "provenance",
          "type"
        ],
        "title": "DerivedEvidence",
        "type": "object"
      },
...
      "EdgeDetail": {
        "additionalProperties": false,
        "properties": {
          "ambiguity_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ambiguity Group"
          },
          "attributes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportsAttributes"
              },
              {
                "$ref": "#/components/schemas/RequiresAttributes"
              },
              {
                "$ref": "#/components/schemas/InheritsAttributes"
              },
              {
                "$ref": "#/components/schemas/HandlesAttributes"
              },
              {
                "$ref": "#/components/schemas/TestsAttributes"
              },
              {
                "$ref": "#/components/schemas/CallsAttributes"
              }
            ],
            "title": "Attributes"
          },
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "contested": {
            "title": "Contested",
            "type": "boolean"
          },
          "evidence": {
            "items": {
              "discriminator": {
                "mapping": {
                  "config": "#/components/schemas/ConfigEvidence",
                  "convention": "#/components/schemas/ConventionEvidence",
                  "derived": "#/components/schemas/DerivedEvidence",
                  "span": "#/components/schemas/SpanEvidence"
                },
                "propertyName": "type"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/SpanEvidence"
                },
                {
                  "$ref": "#/components/schemas/ConfigEvidence"
                },
                {
                  "$ref": "#/components/schemas/ConventionEvidence"
                },
                {
                  "$ref": "#/components/schemas/DerivedEvidence"
                }
              ]
            },
            "title": "Evidence",
            "type": "array"
          },
          "kind": {
            "$ref": "#/components/schemas/EdgeKind"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/NodeSummary"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "target": {
            "$ref": "#/components/schemas/NodeSummary"
          },
          "target_key": {
            "title": "Target Key",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "source_key",
          "target_key",
          "basis",
          "confidence",
          "occurrence_count",
          "ambiguity_group",
          "contested",
          "attributes",
          "source",
          "target",
          "evidence"
        ],
        "title": "EdgeDetail",
        "type": "object"
      },
      "EdgeKind": {
        "description": "Stored relationship kinds. Direction is always ``source -> target``.",
        "enum": [
          "IMPORTS",
          "REQUIRES",
          "INHERITS",
          "HANDLES",
          "TESTS",
          "CALLS"
        ],
        "title": "EdgeKind",
        "type": "string"
      },
      "EdgeRef": {
        "additionalProperties": false,
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/EdgeKind"
          },
          "ref": {
            "const": "edge",
            "title": "Ref",
            "type": "string"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "target_key": {
            "title": "Target Key",
            "type": "string"
          }
        },
        "required": [
          "ref",
          "kind",
          "source_key",
          "target_key"
        ],
        "title": "EdgeRef",
        "type": "object"
      },
...
      "ExtractionStatus": {
        "description": "Per-file coverage. Absence of a relationship only means something for ``full``.",
        "enum": [
          "full",
          "partial",
          "failed",
          "unsupported",
          "excluded"
        ],
        "title": "ExtractionStatus",
        "type": "string"
      },
...
      "HandlesAttributes": {
        "additionalProperties": false,
        "properties": {
          "framework": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Framework"
          }
        },
        "title": "HandlesAttributes",
        "type": "object"
      },
...
      "ImportsAttributes": {
        "additionalProperties": false,
        "properties": {
          "imported_names": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Imported Names",
            "type": "array"
          },
          "is_dynamic": {
            "default": false,
            "title": "Is Dynamic",
            "type": "boolean"
          },
          "is_reexport": {
            "default": false,
            "title": "Is Reexport",
            "type": "boolean"
          },
          "is_type_only": {
            "default": false,
            "title": "Is Type Only",
            "type": "boolean"
          },
          "specifiers": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Specifiers",
            "type": "array"
          },
          "via": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Via",
            "type": "array"
          }
        },
        "required": [
          "specifiers"
        ],
        "title": "ImportsAttributes",
        "type": "object"
      },
      "InheritanceMode": {
        "enum": [
          "extends",
          "implements",
          "mixin"
        ],
        "title": "InheritanceMode",
        "type": "string"
      },
      "InheritsAttributes": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/InheritanceMode"
          },
          "target_symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Symbol"
          }
        },
        "required": [
          "mode"
        ],
        "title": "InheritsAttributes",
        "type": "object"
      },
...
      "NodeBrief": {
        "additionalProperties": false,
        "description": "Enough to name and link to a node.",
        "properties": {
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "repository",
              "directory",
              "file",
              "type",
              "function",
              "endpoint",
              "package",
              "external_package"
            ],
            "title": "SnapshotNodeKind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "key",
          "kind",
          "name"
        ],
        "title": "NodeBrief",
        "type": "object"
      },
...
      "NodeSummary": {
        "additionalProperties": false,
        "description": "A node as a tree, list or search item.",
        "properties": {
          "child_count": {
            "title": "Child Count",
            "type": "integer"
          },
          "extraction_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "enum": [
              "repository",
              "directory",
              "file",
              "type",
              "function",
              "endpoint",
              "package",
              "external_package"
            ],
            "title": "SnapshotNodeKind",
            "type": "string"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "parent_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Key"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/Tag"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "key",
          "kind",
          "name",
          "parent_key",
          "language",
          "extraction_status",
          "child_count",
          "tags"
        ],
        "title": "NodeSummary",
        "type": "object"
      },
...
      "Page_Relationship_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Relationship"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque; pass as `cursor` to get the next page. Null on the last page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[Relationship]",
        "type": "object"
      },
...
      "Problem": {
        "additionalProperties": false,
        "description": "An error, as RFC 9457 problem details (``application/problem+json``).",
        "properties": {
          "code": {
            "description": "Stable machine-readable code, e.g. node_not_found",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-parameter details of a validation_error",
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "description": "URN identifying the problem type: urn:xraymyrepo:problem:<code>",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "code"
        ],
        "title": "Problem",
        "type": "object"
      },
...
      "Provenance": {
        "additionalProperties": false,
        "description": "Who asserted a claim, and by which rule.\n\n``rule`` is what answers \"why does this exist?\". It is mandatory for\nheuristic and inferred claims and for convention evidence (enforced by the\nevidence and record models), and recommended everywhere else.",
        "properties": {
          "producer": {
            "title": "Producer",
            "type": "string"
          },
          "rule": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rule"
          }
        },
        "required": [
          "producer"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "Relationship": {
        "additionalProperties": false,
        "description": "A stored edge touching the requested node, always in its stored direction\n(``source_key`` -> ``target_key``); inverse edges are never synthesized.",
        "properties": {
          "ambiguity_group": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ambiguity Group"
          },
          "attributes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ImportsAttributes"
              },
              {
                "$ref": "#/components/schemas/RequiresAttributes"
              },
              {
                "$ref": "#/components/schemas/InheritsAttributes"
              },
              {
                "$ref": "#/components/schemas/HandlesAttributes"
              },
              {
                "$ref": "#/components/schemas/TestsAttributes"
              },
              {
                "$ref": "#/components/schemas/CallsAttributes"
              }
            ],
            "title": "Attributes"
          },
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "confidence": {
            "$ref": "#/components/schemas/Confidence"
          },
          "contested": {
            "title": "Contested",
            "type": "boolean"
          },
          "direction": {
            "description": "`out`: the requested node is the source; `in`: it is the target.",
            "enum": [
              "out",
              "in"
            ],
            "title": "Direction",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/EdgeKind"
          },
          "occurrence_count": {
            "title": "Occurrence Count",
            "type": "integer"
          },
          "peer": {
            "$ref": "#/components/schemas/NodeBrief"
          },
          "source_key": {
            "title": "Source Key",
            "type": "string"
          },
          "target_key": {
            "title": "Target Key",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "source_key",
          "target_key",
          "basis",
          "confidence",
          "occurrence_count",
          "ambiguity_group",
          "contested",
          "attributes",
          "direction",
          "peer"
        ],
        "title": "Relationship",
        "type": "object"
      },
...
      "RequirementScope": {
        "enum": [
          "prod",
          "dev",
          "optional",
          "peer"
        ],
        "title": "RequirementScope",
        "type": "string"
      },
      "RequiresAttributes": {
        "additionalProperties": false,
        "properties": {
          "scope": {
            "$ref": "#/components/schemas/RequirementScope"
          },
          "version_constraint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version Constraint"
          }
        },
        "required": [
          "scope"
        ],
        "title": "RequiresAttributes",
        "type": "object"
      },
...
      "SpanEvidence": {
        "additionalProperties": false,
        "description": "A location in source code, e.g. an import statement or a decorator.",
        "properties": {
          "ast_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ast Type"
          },
          "basis": {
            "$ref": "#/components/schemas/Basis"
          },
          "end_col": {
            "minimum": 0.0,
            "title": "End Col",
            "type": "integer"
          },
          "end_line": {
            "minimum": 1.0,
            "title": "End Line",
            "type": "integer"
          },
          "file_key": {
            "title": "File Key",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "start_col": {
            "minimum": 0.0,
            "title": "Start Col",
            "type": "integer"
          },
          "start_line": {
            "minimum": 1.0,
            "title": "Start Line",
            "type": "integer"
          },
          "type": {
            "const": "span",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "start_line",
          "start_col",
          "end_line",
          "end_col",
          "basis",
          "provenance",
          "type",
          "file_key"
        ],
        "title": "SpanEvidence",
        "type": "object"
      },
      "Tag": {
        "additionalProperties": false,
        "description": "A classification reduced to facet and value, for badges and colouring.",
        "properties": {
          "facet": {
            "$ref": "#/components/schemas/ClassificationFacet"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "facet",
          "value"
        ],
        "title": "Tag",
        "type": "object"
      },
      "TestsAttributes": {
        "additionalProperties": false,
        "properties": {},
        "title": "TestsAttributes",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Read-only access to persisted Codebase Intelligence Model snapshots.",
    "title": "XRayMyRepo API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/snapshots/{snapshot_id}/edges/by-ref": {
      "get": {
        "operationId": "get_edge_api_v1_snapshots__snapshot_id__edges_by_ref_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EdgeKind"
            }
          },
          {
            "description": "Source node key",
            "in": "query",
            "name": "source",
            "required": true,
            "schema": {
              "description": "Source node key",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "description": "Target node key",
            "in": "query",
            "name": "target",
            "required": true,
            "schema": {
              "description": "Target node key",
              "title": "Target",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EdgeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "One edge, identified by kind, source key and target key, with its evidence",
        "tags": [
          "relationships"
        ]
      }
    },
...
    "/api/v1/snapshots/{snapshot_id}/relationships": {
      "get": {
        "description": "Outgoing edges first, then incoming, each by kind and peer key. Every item keeps\nthe edge's stored direction; ``direction`` only says which end ``node`` is.",
        "operationId": "relationships_api_v1_snapshots__snapshot_id__relationships_get",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "CIM node key",
            "in": "query",
            "name": "node",
            "required": true,
            "schema": {
              "description": "CIM node key",
              "title": "Node",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "direction",
            "required": false,
            "schema": {
              "default": "both",
              "enum": [
                "out",
                "in",
                "both"
              ],
              "title": "Direction",
              "type": "string"
            }
          },
          {
            "description": "Repeatable",
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/EdgeKind"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Repeatable",
              "title": "Kind"
            }
          },
          {
            "description": "Page size",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 200,
              "description": "Page size",
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`next_cursor` of the previous page",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`next_cursor` of the previous page",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Relationship_"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid cursor"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Snapshot not complete"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Invalid parameters or key"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              },
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Database unavailable"
          }
        },
        "summary": "Stored edges that start or end at a node",
        "tags": [
          "relationships"
        ]
      }
    },
...
  }
}